Split types out of ProjectBuilding to avoid cycles
We're going to need ProjectPlanOutput to import the ProjectBuilding
types, but this would cause a cycle with ProjectPlanning. So do the same
thing we did with ProjectPlanning and move the types out.
